refactor: 메인 페이지 컴포넌트를 기능별 폴더로 재구조화
- liveBattles 폴더에 라이브 배틀 관련 컴포넌트 그룹화
- pastBattles 폴더에 종료된 배틀 관련 컴포넌트 그룹화
