+++ frontend/src/pages/mainPage/components/liveBattles/LiveBattlesSection.tsx
