+++ frontend/src/pages/mainPage/components/pastBattles/PastBattlesSection.tsx
